api/requirements.txt
Django>=3.0,<=4.1
psycopg2>=2.9
djangorestframework==3.13.1
django-filter